Stop crawling to UniProt and BioPortal
Obtained from:	https://github.com/level-five/umakadata/issues/386#issuecomment-265079421
Obtained from:	https://github.com/level-five/umakadata/issues/409#issue-195133141

+++ web/lib/tasks/umakadata.rake
@@ -114,6 +114,8 @@ namespace :umakadata do
     end
     rdf_prefixes = RdfPrefix.all.pluck(:id, :endpoint_id, :uri)
     Endpoint.all.order("id #{args[:order]}").each do |endpoint|
+      next if endpoint.id == 56  # UniProt
+      next if endpoint.id == 115 # BioPortal
       crawl_log.evaluations.where(endpoint_id: endpoint.id).delete_all
 
       rdf_prefixes_candidates = Array.new
